# Conflicts:
#	.idea/deploymentTargetDropDown.xml

+++ app/src/main/java/it/uniba/dib/sms222332/professor/ThesisDescriptionProfessorFragment.java
@@ -24,44 +24,36 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.material.snackbar.Snackbar;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.storage.FirebaseStorage;
-import com.google.firebase.storage.ListResult;
 import com.google.firebase.storage.StorageReference;
 
+import java.util.Objects;
+
 import it.uniba.dib.sms222332.R;
+import it.uniba.dib.sms222332.commonActivities.MainActivity;
 
 public class ThesisDescriptionProfessorFragment extends Fragment {
 
-    TextView txtNameTitle,txtType,txtDepartment, txtTime,txtCorrelator,
-            txtDescription,txtRelatedProjects,txtAverageMarks, txtRequiredExams,txtStudentTitle,txtStudent;
+    TextView txtThesisName,txtType,txtDepartment, txtTime,txtCorrelator,
+            txtDescription,txtRelatedProjects,txtAverageMarks, txtRequiredExams,txtStudent;
     Button btnEdit,btnDelete,btnReceipt,btnTask;
     FirebaseFirestore db = FirebaseFirestore.getInstance();
+    StorageReference storageReference = FirebaseStorage.getInstance().getReference();
     LinearLayout layout_lista_file;
-    String related_projects = "" ;
-    String average_marks = "" ;
-    String required_exam = "";
-    String student = "";
-    StorageReference storageReference, ref;
-    FirebaseStorage storage;
-    String nameThesis = "";
-
-
-
+    String relator;
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        ((AppCompatActivity)getActivity()).getSupportActionBar().setTitle(getResources().getString(R.string.thesisListToolbar));
+        Objects.requireNonNull(( (AppCompatActivity) requireActivity() ).getSupportActionBar()).setTitle(getResources().getString(R.string.thesisListToolbar));
 
         View view = inflater.inflate(R.layout.fragment_thesis_description_professor, container, false);
 
         layout_lista_file = view.findViewById(R.id.layout_lista_file);
-        txtNameTitle = view.findViewById(R.id.txtNameTitle);
+        txtThesisName = view.findViewById(R.id.txtNameTitle);
         txtDepartment = view.findViewById(R.id.txtDepartment);
         txtType = view.findViewById(R.id.txtTypology);
         txtTime = view.findViewById(R.id.txtTime);
@@ -70,30 +62,26 @@ public class ThesisDescriptionProfessorFragment extends Fragment {
         txtRelatedProjects = view.findViewById(R.id.txtRelatedProjects);
         txtAverageMarks = view.findViewById(R.id.txtAverageMarks);
         txtRequiredExams = view.findViewById(R.id.txtRequiredExams);
-        btnEdit = view.findViewById(R.id.btnModify);
-        btnDelete = view.findViewById(R.id.btnDelete);
-        txtStudentTitle = view.findViewById(R.id.txtStudentTitle);
+        btnEdit = view.findViewById(R.id.btnEditThesis);
+        btnDelete = view.findViewById(R.id.btnDeleteThesis);
         txtStudent = view.findViewById(R.id.txtStudent);
         btnReceipt = view.findViewById(R.id.btnReceipt);
         btnTask = view.findViewById(R.id.btnTask);
-
-        storage = FirebaseStorage.getInstance();
-        storageReference = storage.getReference();
-
 
         if (getArguments() != null) {
             String correlator = getArguments().getString("correlator");
             String description = getArguments().getString("description");
             String estimated_time = getArguments().getString("estimated_time");
             String faculty = getArguments().getString("faculty");
-            nameThesis = getArguments().getString("name");
+            String thesisName = getArguments().getString("name");
             String type = getArguments().getString("type");
-            related_projects = getArguments().getString("related_projects");
-            average_marks = getArguments().getString("average_marks");
-            required_exam = getArguments().getString("required_exam");
-            student = getArguments().getString("student");
+            String relatedProjects = getArguments().getString("related_projects");
+            String averageMarks = getArguments().getString("average_marks");
+            String requiredExam = getArguments().getString("required_exam");
+            String student = getArguments().getString("student");
+            relator = getArguments().getString("professor");
 
-            txtNameTitle.setText(nameThesis);
+            txtThesisName.setText(thesisName);
             txtType.setText(type);
             txtDepartment.setText(faculty);
             txtTime.setText(estimated_time);
@@ -101,33 +89,32 @@ public class ThesisDescriptionProfessorFragment extends Fragment {
 
 
             if(correlator.isEmpty())
-               txtCorrelator.setText("None");
+               txtCorrelator.setText(R.string.none );
             else
                 txtCorrelator.setText(correlator);
 
-            if(student.isEmpty()){
-                txtStudent.setText("None");
-            }else
+            if(student.isEmpty())
+                txtStudent.setText(R.string.none);
+            else
                 txtStudent.setText(student);
 
-            if(average_marks.isEmpty()){
-                txtAverageMarks.setText("None");
-            }else
-                txtAverageMarks.setText(average_marks);
+            if(averageMarks.isEmpty())
+                txtAverageMarks.setText(R.string.none);
+            else
+                txtAverageMarks.setText(averageMarks);
 
-            if(required_exam.isEmpty()){
-                txtRequiredExams.setText("None");
-            }else
-                txtRequiredExams.setText(required_exam);
+            if(requiredExam.isEmpty())
+                txtRequiredExams.setText(R.string.none);
+            else
+                txtRequiredExams.setText(requiredExam);
 
-            if(related_projects.isEmpty()){
-                txtRelatedProjects.setText("None");
-            }else
-                txtRelatedProjects.setText(related_projects);
-
+            if(relatedProjects.isEmpty())
+                txtRelatedProjects.setText(R.string.none);
+            else
+                txtRelatedProjects.setText(relatedProjects);
         }
 
-        if(!student.isEmpty()){
+        if(!txtStudent.getText().toString().equals(getResources().getString(R.string.none))){
             btnTask.setVisibility(View.VISIBLE);
             btnReceipt.setVisibility(View.VISIBLE);
         }
@@ -136,16 +123,16 @@ public class ThesisDescriptionProfessorFragment extends Fragment {
             Fragment editThesisFragment = new EditThesisFragment();
             Bundle bundle = new Bundle();
 
-            bundle.putString("name",txtNameTitle.getText().toString());
+            bundle.putString("name", txtThesisName.getText().toString());
             bundle.putString("type",txtType.getText().toString());
-            bundle.putString("related_projects",related_projects);
+            bundle.putString("related_projects",txtRelatedProjects.getText().toString());
             bundle.putString("department",txtDepartment.getText().toString());
             bundle.putString("time",txtTime.getText().toString());
             bundle.putString("correlator",txtCorrelator.getText().toString());
             bundle.putString("description",txtDescription.getText().toString());
             bundle.putString("student",txtStudent.getText().toString());
-            bundle.putString("required_exam",required_exam);
-            bundle.putString("average_marks",average_marks);
+            bundle.putString("required_exam",txtRequiredExams.getText().toString());
+            bundle.putString("average_marks", txtAverageMarks.getText().toString());
 
             editThesisFragment.setArguments(bundle);
 
@@ -156,59 +143,61 @@ public class ThesisDescriptionProfessorFragment extends Fragment {
             fragmentTransaction.commit();
         });
 
-        btnDelete.setOnClickListener(view12 -> {
+        if(txtCorrelator.getText().toString().equals(MainActivity.account.getName() + " " + MainActivity.account.getSurname()))
+            btnDelete.setVisibility(View.INVISIBLE);
 
-            DocumentReference tesi = db.collection("Tesi").document(txtNameTitle.getText().toString());
-            FirebaseStorage storage = FirebaseStorage.getInstance();
-            StorageReference storageRef = storage.getReference();
-            StorageReference folderRef = storageRef.child(txtNameTitle.getText().toString());
+        else{
+            btnDelete.setOnClickListener(view12 -> {
+
+                DocumentReference tesi = db.collection("Tesi").document(txtThesisName.getText().toString());
+                FirebaseStorage storage = FirebaseStorage.getInstance();
+                StorageReference storageRef = storage.getReference();
+                StorageReference folderRef = storageRef.child(txtThesisName.getText().toString());
 
 
-            AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-            builder.setTitle("Conferma eliminazione");
-            builder.setMessage("Sei sicuro di voler eliminare questa tesi ed i suoi materiali?");
-            builder.setPositiveButton("No", (dialog, which) -> {
+                AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
+                builder.setTitle(R.string.confirm_deletion);
+                builder.setMessage(R.string.thesis_materials_delete_question);
+                builder.setPositiveButton(R.string.yes, (dialog, which) -> {
 
-            });
-
-            builder.setNegativeButton("Yes", (dialog, which) -> {
-
-                folderRef.listAll()
-                        .addOnSuccessListener(new OnSuccessListener<ListResult>() {
-                            @Override
-                            public void onSuccess(ListResult listResult) {
+                    folderRef.listAll()
+                            .addOnSuccessListener(listResult -> {
                                 for (StorageReference item : listResult.getItems()) {
                                     item.delete();
                                 }
                                 folderRef.delete();
-                            }
-                        })
-                        .addOnFailureListener(new OnFailureListener() {
-                            @Override
-                            public void onFailure(@NonNull Exception e) {
-                                Log.e(TAG, "Error deleting folder: " + e.getMessage()); //TODO INSERIRE MESSAGGIO ERRORE
-                            }
-                        });
+                            })
+                            .addOnFailureListener(e -> Log.e(TAG, R.string.error_deleting_folder + e.getMessage()));
 
-                storageReference.child("PDF_tesi").child(nameThesis + ".pdf").delete();
-                tesi.delete();
+                    storageReference.child("PDF_tesi").child(txtThesisName.getText().toString() + ".pdf").delete();
+                    tesi.delete();
 
-                Snackbar.make(view12, "Thesis eliminated", Snackbar.LENGTH_LONG).show();
 
-                FragmentTransaction transaction = getParentFragmentManager().beginTransaction();
-                transaction.replace(R.id.fragment_container, new ThesesListFragment());
-                transaction.commit();
+                    Snackbar.make(view12, R.string.thesis_eliminated, Snackbar.LENGTH_LONG).show();
+
+                    FragmentTransaction transaction = getParentFragmentManager().beginTransaction();
+                    transaction.replace(R.id.fragment_container, new ThesesListFragment());
+                    transaction.commit();
+
+                });
+
+                builder.setNegativeButton(R.string.no, (dialog, which) -> {
+
+                });
+
+                AlertDialog dialog = builder.create();
+                dialog.show();
 
             });
+        }
 
-            AlertDialog dialog = builder.create();
-            dialog.show();
 
-        });
+
+
 
         //AGGIUNGO MATERIALI DEL DATABASE
         FirebaseStorage storage = FirebaseStorage.getInstance();
-        StorageReference storageRef = storage.getReference().child(txtNameTitle.getText().toString());
+        StorageReference storageRef = storage.getReference().child(txtThesisName.getText().toString());
 
         storageRef.listAll().addOnSuccessListener(listResult -> {
 
@@ -223,8 +212,14 @@ public class ThesisDescriptionProfessorFragment extends Fragment {
         btnReceipt.setOnClickListener(v -> {
 
             Bundle bundle = new Bundle();
-            bundle.putString("thesis_name",txtNameTitle.getText().toString());
+            bundle.putString("thesis_name", txtThesisName.getText().toString());
             bundle.putString("student",txtStudent.getText().toString());
+
+            if(relator.equals(MainActivity.account.getEmail()))
+                bundle.putString("professor", "");
+            else
+                bundle.putString("professor", relator);
+
             Fragment receiptsListFragment = new ReceiptsListFragment();
             receiptsListFragment.setArguments(bundle);
 
@@ -242,8 +237,13 @@ public class ThesisDescriptionProfessorFragment extends Fragment {
             Fragment taskListFragment = new TaskListFragment();
             Bundle bundle = new Bundle();
 
-            bundle.putString("thesisName",txtNameTitle.getText().toString());
+            bundle.putString("thesisName", txtThesisName.getText().toString());
             bundle.putString("student",txtStudent.getText().toString());
+
+            if(relator.equals(MainActivity.account.getEmail()))
+                bundle.putString("professor", "");
+            else
+                bundle.putString("professor", relator);
 
             taskListFragment.setArguments(bundle);
 
@@ -270,7 +270,7 @@ public class ThesisDescriptionProfessorFragment extends Fragment {
 
         downloadMaterial.setOnClickListener(view1 -> {
             download(nomeFile);
-            Snackbar.make(view1, "Downloading "+nomeFile, Snackbar.LENGTH_LONG).show();
+            Snackbar.make(view1, R.string.downloading + nomeFile, Snackbar.LENGTH_LONG).show();
         });
 
         layout_lista_file.addView(view);
@@ -278,24 +278,22 @@ public class ThesisDescriptionProfessorFragment extends Fragment {
 
     private void download(String nomeFile) {
 
-        storageReference = FirebaseStorage.getInstance().getReference();
-        ref = storageReference.child(txtNameTitle.getText().toString()).child(nomeFile);
+        StorageReference ref = storageReference.child(txtThesisName.getText().toString()).child(nomeFile);
 
         ref.getDownloadUrl().addOnSuccessListener(uri -> {
             String url  = uri.toString();
-                downloadFile(getActivity(), nomeFile,"", DIRECTORY_DOWNLOADS,url );
+            downloadFile(requireActivity(), nomeFile, DIRECTORY_DOWNLOADS,url );
 
-        }).addOnFailureListener(e -> {
-      });
+        });
     }
 
-    private void downloadFile(Context context, String nomeFile, String fileExtension, String destinationDirectory, String url ) {
+    private void downloadFile(Context context, String nomeFile, String destinationDirectory, String url) {
          DownloadManager downloadManager = (DownloadManager) context.getSystemService(DOWNLOAD_SERVICE);
          Uri uri = Uri.parse(url);
          DownloadManager.Request request = new DownloadManager.Request(uri);
 
          request.setNotificationVisibility(DownloadManager.Request.VISIBILITY_VISIBLE_NOTIFY_COMPLETED);
-         request.setDestinationInExternalFilesDir(context, destinationDirectory, nomeFile + fileExtension);
+         request.setDestinationInExternalFilesDir(context, destinationDirectory, nomeFile + "");
          downloadManager.enqueue(request);
     }
 }
